refactor: 공통 함수 제거

+++ src/main/java/org/example/DeliveryChargeCalculator.java
@@ -1,5 +1,5 @@
 package org.example;
 
 public interface DeliveryChargeCalculator {
-    int getDeliveryPrice(int weight, int price);
+    int getDeliveryPrice(double weight, int price);
 }
